requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
apscheduler==3.10.4
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
aiosqlite==0.20.0